+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG for Sulu Product Bundle
 * feature/addons
     * ENHANCEMENT Added API for product addons
     * ENHANCEMENT Added GET-api for currencies
+    * ENHANCEMENT Added flag for recurring price
 
 * 0.10.6 (2016-06-23)
     * BUGFIX Fixed issue with tags when adding a new product